Set monitor arg in glfwCreateWindow when fullscreen
for both borderless fullscreen and exclusive fullscreen, as mentioned in
the GLFW documentation.

Also made the GLFW_DECORATED window hint depend only on the user
preferences, not on window type, since this hint is ignored for
fullscreen windows.

When creating a fullscreen window of the same size as the monitor's,
GLFW always creates a borderless fullscreen window. As a consequence,
choosing fullscreen with the desktop resolution in the settings has
the same effect as choosing borderless (but takes away the ability to
scale the UI, besides selecting a smaller res which results in blurry
graphics).

See https://discourse.glfw.org/t/actively-choosing-between-fullscreen-and-windowed-fullscreen/2553/3

According to this same source, the bit depths do not matter. I also
tried to request bit depths of 4 and 16, but always got a context
created with 8 bits per channel.